+++ app/src/androidTest/java/com/example/mypet/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.android.mypet
+package com.example.mypet
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
