--- app/src/main/java/com/android/mypet/DBManager.kt
@@ -1,21 +1,0 @@
-package com.android.mypet
-
-import android.content.Context
-import android.database.sqlite.SQLiteDatabase
-import android.database.sqlite.SQLiteOpenHelper
-
-// DBManager 구성
-class DBManager(
-    context: Context?,
-    name: String?,
-    factory: SQLiteDatabase.CursorFactory?,
-    version: Int
-) : SQLiteOpenHelper(context, name, factory, version) {
-    override fun onCreate(db: SQLiteDatabase?) {
-        db!!.execSQL("CREATE TABLE petinfo (종 text, 이름 text, 생일 text, 처음만난날짜 text, 몸무게 text, 질병 text)")
-    }
-
-    override fun onUpgrade(p0: SQLiteDatabase?, p1: Int, p2: Int) {
-        // 갱신하는 거 필요 없어서 추가 작성x
-    }
-}
