+++ app/src/main/java/com/example/mypet/MainActivity.kt
@@ -1,11 +1,12 @@
-package com.android.mypet
+package com.example.mypet
 
+import android.content.Context
 import android.content.Intent
+import android.database.sqlite.SQLiteDatabase
+import android.database.sqlite.SQLiteOpenHelper
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
-import com.example.mypet.MyPage
-import com.example.mypet.R
 
 class MainActivity : AppCompatActivity() {
 
@@ -25,11 +26,5 @@ class MainActivity : AppCompatActivity() {
             val nextIntent = Intent(this, PetRegistration::class.java)
             startActivity(nextIntent)
         }
-        var myPageGo = findViewById<Button>(R.id.myPageB)
-        myPageGo.setOnClickListener{
-            var mypageIntent = Intent(this, MyPage::class.java)
-            startActivity(mypageIntent)
-        }
-
     }
 }
