+++ app/src/main/java/com/example/mypet/SignUp_sub2.kt
@@ -1,0 +1,11 @@
+package com.example.mypet
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+
+class SignUp_sub2 : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_sign_up_sub2)
+    }
+}
