+++ app/src/main/res/drawable/shadow.xml
@@ -5,9 +5,5 @@
     <item> <shape> <padding android:top="0dp" android:right="2dp" android:bottom="2dp" android:left="0dp" /> <corners android:radius="10dp" /> <solid android:color="#8DDBDBDB" /> </shape> </item>
     <item> <shape> <padding android:top="0dp" android:right="2dp" android:bottom="2dp" android:left="0dp" /> <corners android:radius="10dp" /> <solid android:color="#57DFDFDF" /> </shape> </item>
     <!-- Background -->
-<<<<<<< HEAD
     <item> <shape> <solid android:color="#F8F8F8" /> <corners android:radius="10dp" /> </shape> </item>
-=======
-    <item> <shape> <solid android:color="#BBBBBB" /> <corners android:radius="10dp" /> </shape> </item>
->>>>>>> main
 </layer-list>
