+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,6 @@
     android:clipToOutline="true"
     >
 
-
     <Button
         android:id="@+id/button5"
         android:layout_width="wrap_content"
@@ -18,15 +17,14 @@
         android:layout_marginEnd="157dp"
         android:text="임시" />
 
-
     <Button
         android:id="@+id/button6"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_below="@+id/button5"
         android:layout_alignParentEnd="true"
-        android:layout_marginTop="67dp"
-        android:layout_marginEnd="104dp"
-        android:text="반려동물 등록 화면 임시" />
+        android:layout_marginTop="62dp"
+        android:layout_marginEnd="154dp"
+        android:text="Button" />
 </RelativeLayout>
