+++ app/src/main/res/layout/activity_pet_registration.xml
@@ -322,4 +322,3 @@
 
 
 </LinearLayout>
-
