+++ app/src/main/res/layout/activity_sign_up_sub2.xml
@@ -11,7 +11,11 @@
         android:id="@+id/signUpNextButton"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginBottom="328dp"
+        android:background="@drawable/buttonborder"
+        android:backgroundTint="#FF913F"
+        android:text="다 음"
+        android:textSize="16sp"
+        android:textStyle="bold"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.492"
         app:layout_constraintStart_toStartOf="parent"
@@ -33,18 +37,16 @@
         app:layout_constraintStart_toStartOf="parent"
         tools:layout_editor_absoluteY="144dp" />
 
-
     <TextView
         android:id="@+id/welcomeT"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="28dp"
-        android:layout_marginTop="5dp"
-        android:layout_marginEnd="173dp"
-        android:layout_marginBottom="1dp"
+        android:text="마이펫에 오신것을 환영합니다! "
+        android:textColor="#949494"
+        android:textSize="16sp"
+        android:textStyle="bold"
         tools:layout_editor_absoluteX="28dp"
         tools:layout_editor_absoluteY="72dp" />
-
 
     <TextView
         android:id="@+id/signUpT"
@@ -52,16 +54,11 @@
         android:layout_height="wrap_content"
         android:layout_marginStart="28dp"
         android:layout_marginTop="32dp"
-
-        android:layout_marginEnd="148dp"
-        android:layout_marginBottom="5dp"
         android:fontFamily="sans-serif"
         android:text="아이디를 입력해주세요!"
         android:textColor="#000000"
         android:textSize="24sp"
         android:textStyle="bold"
-        app:layout_constraintBottom_toTopOf="@+id/welcomeT"
-        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 </androidx.constraintlayout.widget.ConstraintLayout>
